Encapsulate out audio setup. Don't setup audio until required.
